% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{magoutcomes}
\alias{magoutcomes}
\title{Magistrates court outcomes.}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 435402 rows and 21 columns.}
\source{
\url{https://www.gov.uk/government/statistics/criminal-justice-system-statistics-quarterly-december-2018}
}
\usage{
magoutcomes
}
\description{
Outcomes for all defendants appearing at magistrates courts in England and Wales in 2018, published by the Ministry of Justice.
}
\details{
Names have been cleaned for easy use with R, and additional fields for the region and country of the relevant police force added.
}
\keyword{datasets}
